pyproject.toml
[project]
name = "fiap-ml"
version = "0.1.0"
description = ""
authors = [
    {name = "Erick Muller",email = "erick@em.pro.br"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "duckdb (>=1.3.0,<2.0.0)",
    "duckdb-engine (>=0.17.0,<0.18.0)",
    "python-decouple (>=3.8,<4.0)",
]

[tool.poetry]
packages = [{include = "fiap-ml", from = "src"}]


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.14"
pytest = "^8.4.0"
pandas = "^2.3.0"
jupyterlab = "^4.4.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
serve = "fastapi dev src/fiap-ml/main.py"